pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "listerine"
version = "0.1.2"
description = "Self-hostable grocery list backend"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.36",
  "greenlet>=3.1.1",
  "alembic>=1.13.0",
  "asyncpg>=0.29.0",
  "aiosqlite>=0.20.0",
  "pydantic-settings>=2.6.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "jinja2>=3.1.4",
  "sqladmin>=0.23.0",
  "python-multipart>=0.0.9",
  "email-validator>=2.2.0",
  "itsdangerous>=2.2.0",
  "webauthn>=2.7.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "httpx>=0.27.0",
  "coverage[toml]>=7.6.0",
  "black==25.12.0",
  "flake8>=7.1.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["tests*", "alembic*"]

[tool.pytest.ini_options]
addopts = "-q --cov=app --cov-report=term-missing --cov-fail-under=100"
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["app"]

[tool.black]
line-length = 100
target-version = ["py314"]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [".venv"]
